backend/admin_api/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-firestore==2.13.1
google-cloud-bigquery==3.13.0
pydantic==2.5.0
